chore: existing governance flow tests passing
